Move record and collection transforms into trait
